Updated support for Jac linear operator and added tests
